+++ test_files/main.h
@@ -5,8 +5,8 @@ int _printf(const char *format, ...);
 int _putchar(char c);
 int print_string(char *);
 int print_decimal(int value);
-long long int power(int base, int x);
-long long convertToDecimal(int Number, int base);
+int power(int base, int x);
+int convertToDecimal(int Number, int base);
 int print_binary(int number);
 int print_unsigned(unsigned int number);
