+++ test_files/power.c
@@ -1,6 +1,7 @@
-long long int power(int base, int x){
+int power(int base, int x){
     
     int result=1;
+
     
     for (x; x>0; x--)
     {
